requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1.1
djangorestframework==3.11.1
drf-yasg==1.17.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
openapi-codec==1.3.2
packaging==20.4
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyad==0.6.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.31
websocket-client==0.57.0
whitenoise==5.2.0
